Remove unnecessary boxing.
Idea is reporting this as a minor coding style issue.

+++ src/main/java/nikita/common/util/serializers/noark5v4/hateoas/DocumentDescriptionHateoasSerializer.java
@@ -54,7 +54,7 @@ public class DocumentDescriptionHateoasSerializer
         printTitleAndDescription(jgen, documentDescription);
         if (documentDescription.getDocumentNumber() != null) {
             jgen.writeNumberField(DOCUMENT_DESCRIPTION_DOCUMENT_NUMBER,
-                    documentDescription.getDocumentNumber().intValue());
+                    documentDescription.getDocumentNumber());
         }
         printCreateEntity(jgen, documentDescription);
         if (documentDescription.getAssociationDate() != null) {
